+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix batch state update (when using MongoLite) #75
 - Fix min/max settings for number fields #76
 - Set X-Frame-Options to prevent possible clickjacking via iframe layer
+- Allow to load linked content items in display previews #78
 
 ## 2.3.8 (2023-02-04)